Refactor
Split the massive DrawingScene class into multiple separate classes, alongside a few bug fixes.
