+++ QtPaintTest/QtPaintTest/BaseTool.cpp
@@ -1,0 +1,13 @@
+#include "BaseTool.h"
+#include "DrawingScene.h"
+
+BaseTool::BaseTool() {
+	// Initialize other members if needed
+}
+
+void BaseTool::keyPressEvent(QKeyEvent* event) {
+	// Default implementation does nothing
+}
+void BaseTool::keyReleaseEvent(QKeyEvent* event) {
+	// Default implementation does nothing
+}
