+++ QtPaintTest/QtPaintTest/FillTool.h
@@ -1,0 +1,20 @@
+#pragma once
+#include "BaseTool.h"
+
+class DrawingScene;
+
+class FillTool : public BaseTool {
+public:
+	FillTool();
+	~FillTool() override;
+	void mousePressEvent(QGraphicsSceneMouseEvent* event) override;
+	void mouseMoveEvent(QGraphicsSceneMouseEvent* event) override;
+	void mouseReleaseEvent(QGraphicsSceneMouseEvent* event) override;
+	void keyPressEvent(QKeyEvent* event) override;
+	void keyReleaseEvent(QKeyEvent* event) override;
+	QString toolName() const override { return "Fill"; }
+	QIcon toolIcon() const override { return QIcon(":/icons/fill.png"); }
+
+private:
+	void applyFill(const QPointF& pos);
+};
