+++ QtPaintTest/log.txt
@@ -1,0 +1,17 @@
+Key Pressed: 16777249
+Key Pressed: 67
+Key Pressed: 86
+Pasting clipboard items...
+Key Pressed: 16777249
+Key Pressed: 86
+Pasting clipboard items...
+Key Pressed: 16777249
+Key Pressed: 86
+Pasting clipboard items...
+Key Pressed: 86
+Pasting clipboard items...
+Key Pressed: 86
+Pasting clipboard items...
+Key Pressed: 88
+Key Pressed: 88
+Key Pressed: 88
